fix: ws architecture
